add debug switch for debuging ckcp error

+++ ckcp/openshift_dev_setup.sh
@@ -275,9 +275,20 @@ patches:
         value: $ckcp_route " >>"$ckcp_temp_dir/kustomization.yaml"
 
   echo -n "  - kcp $kcp_version: "
-  kubectl apply -k "$ckcp_temp_dir" >/dev/null 2>&1
+  kubectl apply -k "$ckcp_temp_dir" >/dev/null
   # Check if ckcp pod status is Ready
-  kubectl wait --for=condition=Ready -n $ns pod -l=app=kcp-in-a-pod --timeout=90s >/dev/null
+  local error="false"
+  kubectl wait --for=condition=Ready -n "$ns" pod -l=app=kcp-in-a-pod --timeout=120s >/dev/null 2>&1 || error="true"
+  if [ "$error" == "true" ]; then
+    printf "Debug: print out pod state\n"
+    kubectl -n "$ns" describe pods
+
+    printf "Debug: print out pod logs\n"
+    kubectl -n "$ns" logs "$(kubectl -n "$ns" get pods -l=app=kcp-in-a-pod -o name)"
+
+    exit 1
+  fi
+
   # Clean up kustomize temp dir
   rm -rf "$ckcp_temp_dir"
